+++ paraphrasing-tool/app/__init__.py
@@ -1,19 +1,16 @@
-from flask import Flask
 from flask_cors import CORS
-
+import connexion
 
 def create_app():
-    app = Flask(__name__)
-    CORS(app)
+    # Initialize the Flask app
+    app = connexion.App(__name__, specification_dir='./')
+    app.add_api('../openapi.yaml')
 
-    # Import configuration
-    from .config import Config
-
-    app.config.from_object(Config)
-
-    # Import the main blueprint
-    from .main import main as main_blueprint
-
-    app.register_blueprint(main_blueprint)
+    # Configure CORS to allow requests from specific origins
+    CORS(app.app)
 
     return app
+
+if __name__ == "__main__":
+    app = create_app()
+    app.run(host="0.0.0.0", port=5000)
