+++ paraphrasing-tool/openapi.yaml
@@ -1,4 +1,4 @@
-openapi: 3.1.0
+openapi: 3.0.3
 info:
   title: Paraphrasing Tool API
   version: 1.0.0
