+++ paraphrasing-tool/pyproject.toml
@@ -18,6 +18,8 @@ pydantic = ">=2"
 typing-extensions = ">=4.7.1"
 python-dotenv = "^1.0.1"
 openapi-server = {path = "app/generated/server"}
+connexion = {extras = ["flask"], version = "^3.2.0"}
+uvicorn = "^0.34.0"
 
 
 [tool.poetry.group.dev.dependencies]
