+++ paraphrasing-tool/run.py
@@ -3,4 +3,4 @@ from app import create_app
 app = create_app()
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run()
